Move command detection classes to commandfinder
