+++ src/providers/completer/commandlib/commandfinder.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode'
 import * as fs from 'fs'
 import {latexParser} from 'latex-utensils'
-import {Suggestion, CommandSignatureDuplicationDetector, CommandNameDuplicationDetector} from '../command'
+import {Suggestion} from '../command'
 import type {Extension} from '../../../main'
 import type {ILwCompletionItem} from '../interface'
 
@@ -289,3 +289,50 @@ export class CommandFinder {
     }
 
 }
+
+
+export class CommandSignatureDuplicationDetector {
+    private readonly cmdSignatureList: Set<string> = new Set<string>()
+
+    add(cmd: Suggestion) {
+        this.cmdSignatureList.add(cmd.signatureAsString())
+    }
+
+    has(cmd: Suggestion): boolean {
+        return this.cmdSignatureList.has(cmd.signatureAsString())
+    }
+}
+
+export class CommandNameDuplicationDetector {
+    private readonly cmdSignatureList: Set<string> = new Set<string>()
+
+    constructor(suggestions: Suggestion[] = []) {
+        this.cmdSignatureList = new Set<string>(suggestions.map(s => s.name()))
+    }
+
+    add(cmd: Suggestion): void
+    add(cmdName: string): void
+    add(cmd: any): void {
+        if (cmd instanceof Suggestion) {
+            this.cmdSignatureList.add(cmd.name())
+        } else if (typeof(cmd) === 'string') {
+            this.cmdSignatureList.add(cmd)
+        } else {
+            throw new Error('Unaccepted argument type')
+        }
+    }
+
+    has(cmd: Suggestion): boolean
+    has(cmd: string): boolean
+    has(cmd: any): boolean {
+        if (cmd instanceof Suggestion) {
+            return this.cmdSignatureList.has(cmd.name())
+        } else if (typeof(cmd) === 'string') {
+            return this.cmdSignatureList.has(cmd)
+        } else {
+            throw new Error('Unaccepted argument type')
+        }
+    }
+}
+
+
